AGMRS/requirements.txt
asgiref==3.6.0
Django==4.2
django-adminlte2==0.4.3
django-filter==23.2
django-tables2==2.5.3
sqlparse==0.4.4
